Split, convert react-i18n package to TypeScript
