+++ app/javascript/packages/react-i18n/index.ts
@@ -1,0 +1,3 @@
+export { default as formatHTML } from './format-html';
+export { default as I18nContext } from './i18n-context';
+export { default as useI18n } from './use-i18n';
